pyproject.toml
[tool.poetry]
name = "MENDR"
version = "0.1.0"
description = "Measurement Error in Network Diffusion Datasets for graph Reconstruction"
authors = ["Rachael T. Sexton <rtbs@nist.gov>"]
license = "NIST"
readme = "README.md"
packages = [{include = "mendr"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
networkx = "^3.2.1"
numpy = "^1.26.4"
scipy = "^1.12.0"
cyclopts = "^2.3.2"
csrgraph = {git = "https://github.com/VHRanger/CSRGraph.git"}
jaxtyping = "^0.2.25"
beartype = "^0.17.2"
pyserde = "^0.14.0"
sparse = "^0.15.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"
coconut = {extras = ["kernel"], version = "^3.0.4"}
hypothesis = {extras = ["numpy"], version = "^6.98.12"}
pytest = "^8.0.2"
black = "^24.2.0"
hypothesis-networkx = "^0.3.0"
dvc = "^3.49.0"



[tool.poetry.group.report.dependencies]
autoregistry = "^1.1.2"
affinis = {path = "../affinis"}
dvclive = {extras = ["markdown", "sklearn"], version = "^3.45.0"}

[tool.poetry.scripts]
gmd-nist = 'mendr.__main__:main'
# gmd-nist-rw = 'nist_gmd.main:random-graph-walks'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
